% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectGates.R
\name{detectGates}
\alias{detectGates}
\title{Automatically detect Gates based on the t-scores}
\usage{
detectGates(
  template,
  ts,
  thresh = 5,
  minT = 1,
  tryO = 10,
  nM = 2,
  conc = 3.5,
  tSeed = NA
)
}
\arguments{
\item{template}{A HexTemplate Object}

\item{ts}{A vector of t-scores}

\item{thresh}{A threshold that indicates neighboring hexagons might belong to the same region of interest (gate)}

\item{minT}{A minimum t-score so that a hexagon with this score is regarded as a valid entry point}

\item{tryO}{An integer that determines how many times should be looked for the best configuration}

\item{nM}{An integer specifying what is considered a neighbor}
}
\value{
A sf Object
}
\description{
Automatically detect Gates based on the t-scores
}
