% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineNHex.R
\name{determineNHex}
\alias{determineNHex}
\title{Create a Data.Frame of mean Bray Distances per Resolution}
\usage{
determineNHex(fcs, xChannel, yChannel, seqq = seq(5, 50, 5))
}
\arguments{
\item{fcs}{A flowCore flowSet}

\item{xChannel}{name of the x-channel}

\item{yChannel}{name of the y-channel}

\item{seqq}{a vector of resolutions (xbins) to check}
}
\value{
A data.frame of mean Bray distances
}
\description{
Create a Data.Frame of mean Bray Distances per Resolution
}
