% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{frequencies}
\alias{frequencies}
\title{Get table of Frequencies of every Hexagon's Event for Every Sample}
\usage{
frequencies(template)
}
\arguments{
\item{template}{A \code{\link{HexTemplate}}}
}
\value{
a data.frame
}
\description{
Get table of Frequencies of every Hexagon's Event for Every Sample
}
