% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{getMatrix}
\alias{getMatrix}
\title{Get a Matrix with no Gaps from one Sample of a HexTemplate}
\usage{
getMatrix(template, sample = 1)
}
\arguments{
\item{template}{A \code{\link{HexTemplate}}}

\item{sample}{An integer representing a sample}
}
\value{
A matrix
}
\description{
Get a Matrix with no Gaps from one Sample of a HexTemplate
}
