% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex2Spatial.R
\name{hex2Spatial}
\alias{hex2Spatial}
\title{Convert a Hexbin Object to a SpatialPolygons Object}
\usage{
hex2Spatial(hex)
}
\arguments{
\item{hex}{A Hexbin Object}
}
\value{
A SpatialPolygons Object
}
\description{
Convert a Hexbin Object to a SpatialPolygons Object
}
