% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexTemplate_s4.R
\docType{class}
\name{HexTemplate-class}
\alias{HexTemplate-class}
\title{S4 Class HexTemplate}
\value{

}
\description{
S4 Class HexTemplate
}
\section{Slots}{

\describe{
\item{\code{hexagonSP}}{SpatialPolygons that contain information about the hexagons.}

\item{\code{counts}}{list of events per hexagon per sample.}

\item{\code{metadata}}{data.frame.}

\item{\code{xChannel}}{character.}

\item{\code{yChannel}}{character.}

\item{\code{nHex}}{numeric.}

\item{\code{nSamples}}{numeric.}
}}

