% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoi.R
\name{hoi}
\alias{hoi}
\title{Plot a Grid of t-scores}
\usage{
hoi(template, group = 1)
}
\arguments{
\item{template}{A HexTemplate}

\item{group}{The column name or index of a metadata group}
}
\value{
A list of ggplots
}
\description{
Plot a Grid of t-scores
}
