% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{meta.data}
\alias{meta.data}
\title{Getter-Function for a HexTemplate's Metadata}
\usage{
meta.data(hexTemplate)
}
\arguments{
\item{hexTemplate}{A \code{\link{HexTemplate}}}
}
\value{
a \code{data.frame} (The HexTemplate's metadata slot)
}
\description{
Getter-Function for a HexTemplate's Metadata
}
\examples{
hexT <- data(Testplate)
head(meta.data(hexT))
}
