% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaySP.R
\name{overlaySP}
\alias{overlaySP}
\title{Bins a Data.Frame of Coordinates into a SpatialPolygons Object}
\usage{
overlaySP(hexagonSP, data)
}
\arguments{
\item{hexagonSP}{A \code{\link{SpatialPolygons}} objects}

\item{data}{A \code{data.frame} of coordinates}
}
\value{
A list of binned events per polygon
}
\description{
Bins a Data.Frame of Coordinates into a SpatialPolygons Object
}
