% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHexTemplate.R
\name{plot.HexTemplate}
\alias{plot.HexTemplate}
\title{Plot an empty or filled HexTemplate}
\usage{
\method{plot}{HexTemplate}(hexTemplate, sample = NA, ...)
}
\arguments{
\item{hexTemplate}{A \code{\link{HexTemplate}} Object}

\item{sample}{A positive number corresponding to the sample with which the template should be filled. Leave NA to plot empty template.}

\item{...}{Additional arguments passed to \code{geom_sf}}
}
\value{
A ggplot2 plot
}
\description{
Plot an empty or filled HexTemplate
}
