% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tscores.R
\name{plot_tscores}
\alias{plot_tscores}
\title{Convenience Function for plotting t-scores}
\usage{
plot_tscores(template, ts, ...)
}
\arguments{
\item{template}{A HexTemplate Object}

\item{ts}{A vector of t-scores}

\item{...}{Additional arguments passed to \code{geom_sf}}
}
\value{
A ggplot Object
}
\description{
Convenience Function for plotting t-scores
}
