% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscores.R
\name{tscores}
\alias{tscores}
\title{Get t-scores for a HexTemplate and a Metadata Group}
\usage{
tscores(template, group = 1)
}
\arguments{
\item{template}{A HexTemplate Object}

\item{group}{A metadata group}
}
\value{
A DataFrame of t-scores
}
\description{
Get t-scores for a HexTemplate and a Metadata Group
}
