% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexTemplate.R
\name{HexTemplate}
\alias{HexTemplate}
\title{Title}
\usage{
HexTemplate(
  flowset,
  xChannel,
  yChannel,
  xbins = 20,
  metadata = data.frame(),
  trans_fun = log10
)
}
\arguments{
\item{flowset}{A flowCore \code{flowSet}}

\item{xChannel}{A string representing the x-axis channel}

\item{yChannel}{A string representing the y-axis channel}

\item{xbins}{Number of hexagons on the x-axis - 1}

\item{metadata}{A data.frame containing metadata. Column names should be the group names.}

\item{trans_fun}{A function to transform the data. Standard is \code{log10}. \code{return} if data should not be transformed.}
}
\value{
A HexTemplate
}
\description{
Title
}
\examples{
#get example fcs
#get example metadata
#show available channels
colnames(fcs)
hexT <- HexTemplate(flowset = fcs, xChannel = "SSC-H", yChannel = "DAPI-H", metadata)
}
